// app/types/donation.ts
export type Donation = {
  donation_id: string; // generated by database
  date_submitted: Date | string; // generated by default
  donor_is_individual: boolean;
  donor_individual_name?: string | null; // optional
  donor_business_name?: string | null; // optional
  donor_business_contact_name?: string | null; // optional
  donor_email: string;
  donor_phone?: string | null; // optional
  donor_street_address?: string | null; // optional
  donor_receive_mailings: boolean;
  donor_receive_emails: boolean;
  donor_remain_anonymous: boolean;
  estimated_value: number;
  items_donated: string;
  receiver_first_name: string;
  receiver_last_name: string;
  store_name?: string | null;
  store_street_address?: string | null;
};

export type DonationInsert = Omit<Donation, 'donation_id' | 'date_submitted'>;